rename downloader component to reflect change to cardano
