+++ src/components/MarkdownComponents/index.js
@@ -1,21 +1,19 @@
-import ByronDaedalusDownloaders from './ByronDaedalusDownloaders'
+import CardanoDownloader from './CardanoDownloader'
 import ByronFaucet from './ByronFaucet'
 import IELEFaucet from './IELEFaucet'
 import ITNFaucet from './ITNFaucet'
 import KEVMFaucet from './KEVMFaucet'
 import OtherAvailableTestnets from './OtherAvailableTestnets'
-import ShelleyDaedalusDownloaders from './ShelleyDaedalusDownloaders'
 import ShelleyHaskellFaucet from './ShelleyHaskellFaucet'
 import ShelleyHaskellStakingCalculator from './ShelleyHaskellStakingCalculator'
 
 export default {
-  ByronDaedalusDownloaders,
+  CardanoDownloader,
   ByronFaucet,
   IELEFaucet,
   ITNFaucet,
   KEVMFaucet,
   OtherAvailableTestnets,
-  ShelleyDaedalusDownloaders,
   ShelleyHaskellFaucet,
   ShelleyHaskellStakingCalculator
 }
